Add 'unknown' for an acceptable type for RequestBody
When trying to do a POST or PUT it's mad about trying to send in a typed object in Typescript. Adding unknown means we can send whatever we want to the server. The type safety would be on the function calling it, not on the post or put method of domo.js.